requirements.txt
google-genai
Pillow
python-dotenv
streamlit